refactor AbstractAnthropic to extend Task and remove AbstractAnthropicBase
