+++ src/main/java/io/kestra/plugin/anthropic/AbstractAnthropic.java
@@ -1,9 +1,12 @@
 package io.kestra.plugin.anthropic;
 
 import com.anthropic.models.messages.Message;
+import com.anthropic.client.AnthropicClient;
+import com.anthropic.client.okhttp.AnthropicOkHttpClient;
 
 import io.kestra.core.models.executions.metrics.Counter;
 import io.kestra.core.models.property.Property;
+import io.kestra.core.models.tasks.Task;
 import io.kestra.core.runners.RunContext;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -22,7 +25,12 @@ import io.kestra.core.models.annotations.PluginProperty;
 @EqualsAndHashCode
 @Getter
 @NoArgsConstructor
-public abstract class AbstractAnthropic extends AbstractAnthropicBase {
+public abstract class AbstractAnthropic extends Task {
+
+    @Schema(title = "Anthropic API Key")
+    @NotNull
+    @PluginProperty(secret = true, group = "main")
+    protected Property<String> apiKey;
 
     @Schema(
         title = "Model",
@@ -70,6 +78,12 @@ public abstract class AbstractAnthropic extends AbstractAnthropicBase {
         message.usage().cacheReadInputTokens().ifPresent(tokens ->
             runContext.metric(Counter.of("usage.cache.read.tokens", tokens))
         );
+    }
+
+    protected AnthropicClient buildClient(String rApiKey) {
+        return AnthropicOkHttpClient.builder()
+            .apiKey(rApiKey)
+            .build();
     }
 
 }
